requirements.txt
# --- PDI-Eval: 核心深度学习环境 (锁定以保证可复现性) ---
# 匹配 CUDA 11.8
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# ---  perception: 关键感知模型接口 ---
# 注意：transformers 必须锁定 4.41.2 以兼容 Torch 2.1.0
transformers==4.41.2
accelerate>=0.26.0
timm==1.0.7
einops==0.8.0

# --- 几何与 3D 处理 (Mega-SAM 必需) ---
# 必须匹配 torch-2.1.0+cu118
--find-links https://data.pyg.org/whl/torch-2.1.0+cu118.html
torch-scatter
kornia==0.7.4
scikit-learn
open3d
ninja

# --- 图像与视频 IO ---
# 建议只保留一个 opencv 分发版
opencv-python>=4.8.0
imageio==2.36.0
decord==0.6.0
pillow>=10.0.0

# --- 基础工具与可视化 ---
numpy==1.26.4
scipy==1.14.1
matplotlib
tqdm
pyyaml
rich
gradio

# --- 外部子模块 (安装指令参考) ---
# 建议在 README 中说明手动安装，或保留以下行（需确保 git 环境可用）
# git+https://github.com/facebookresearch/segment-anything-2.git
# git+https://github.com/facebookresearch/co-tracker.git